[hir2mpl]Internal commit msg:     [IPA] code clean.     [hir2mpl] clean up pclint in common, part 2     [hir2mpl] clean up pclint alarm in ast_input [mapleall]Internal commit msg:     [mplcg]Enable ElimSpecificAnd     [IPA] code clean.     [mplcg] prepeephole LongIntCompareWithZPattern use AArch64CGPeepHole     !1053 Added "--calleeregs-placement" opt. Default is off for now.     !1033 [mapleall] Fix bugs related to iterators invalidation

+++ src/hir2mpl/common/include/feir_var.h
@@ -122,7 +122,7 @@ class FEIRVar {
 
   MIRSymbol *GenerateGlobalMIRSymbol(MIRBuilder &builder) const {
     MIRSymbol *mirSym = GenerateGlobalMIRSymbolImpl(builder);
-    mirSym->GetSrcPosition().SetFileNum(srcFileIdx);
+    mirSym->GetSrcPosition().SetFileNum(static_cast<uint16>(srcFileIdx));
     mirSym->GetSrcPosition().SetLineNum(srcFileLineNum);
     builder.GetMirModule().InsertInlineGlobal(mirSym->GetStIdx().Idx());
     return mirSym;
@@ -130,7 +130,7 @@ class FEIRVar {
 
   MIRSymbol *GenerateLocalMIRSymbol(MIRBuilder &builder) const {
     MIRSymbol *mirSym = GenerateLocalMIRSymbolImpl(builder);
-    mirSym->GetSrcPosition().SetFileNum(srcFileIdx);
+    mirSym->GetSrcPosition().SetFileNum(static_cast<uint16>(srcFileIdx));
     mirSym->GetSrcPosition().SetLineNum(srcFileLineNum);
     return mirSym;
   }
